requirements.txt
altair @ file:///home/conda/feedstock_root/build_artifacts/altair-split_1734244716962/work
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1741918516150/work
biopython @ file:///Users/runner/miniforge3/conda-bld/biopython_1737241752615/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1725267563793/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1740094013202/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1746569525376/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1725560577695/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1746951630646/work
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1744743111830/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1746913637690/work
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1735887193964/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1735929639977/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1733472696581/work
jsonschema-specifications @ file:///tmp/tmpuvkyqc9y/src
kiwisolver @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_cc2l_z_0ri/croot/kiwisolver_1737039586949/work
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1733219703897/work
matplotlib==3.10.3
munkres==1.1.4
narwhals @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_narwhals_1747340565/work
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1745119054649/work/dist/numpy-2.2.5-cp313-cp313-macosx_11_0_arm64.whl#sha256=3c7ebe093f9d5df963ac6e957c7ac1cb888f09d41c2c1cd9cf70c063909aed80
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1744430513098/work
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1733792384640/work
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1746646311040/work
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1733344503739/work
protobuf @ file:///Users/runner/miniforge3/conda-bld/protobuf_1741124664418/work/bazel-bin/python/dist/protobuf-5.29.3-cp313-abi3-macosx_11_0_arm64.whl#sha256=1c27ec799bd9d9dfc910617df10de21856e24ac0ae97f3dfb2f6540ffc2d1f4b
pyarrow==20.0.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydeck @ file:///home/conda/feedstock_root/build_artifacts/pydeck_1738346561094/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1743089729650/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1733215673016/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1737454709251/work
referencing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_referencing_1737836872/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rpds-py @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_rpds-py_1747322953/work
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1739790783802/work/dist/scipy-1.15.2-cp313-cp313-macosx_11_0_arm64.whl#sha256=b07c2380f54734e2bc5d19420a293aabf6bcd5a34beba80eace1959b32b5edde
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1733730015268/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1739781697784/work
statsmodels @ file:///Users/runner/miniforge3/conda-bld/statsmodels_1727986717451/work
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1747153904630/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1743640740204/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1747384505886/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1744302253/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1744323578849/work
watchdog @ file:///Users/runner/miniforge3/conda-bld/watchdog_1730492921197/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard==0.23.0
